break up registration files
